+++ huskeliste/app/src/main/java/com/stegemoen/huskeliste/todolists/TodoItemRecyclerAdapter.kt
@@ -3,23 +3,26 @@ package com.stegemoen.huskeliste.todolists
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.CheckBox
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.stegemoen.huskeliste.R
 import com.stegemoen.huskeliste.todolists.data.TodoItem
 
 class TodoItemRecyclerAdapter(
-    private val todoItems: MutableList<TodoItem>) //,
-    //private val onItemClicked:(TodoItem)->Unit)
-    : RecyclerView.Adapter<TodoItemRecyclerAdapter.ViewHolder>() {
+        private val todoItems: MutableList<TodoItem>) //,
+        //private val onItemClicked:(TodoItem)->Unit)
+        : RecyclerView.Adapter<TodoItemRecyclerAdapter.ViewHolder>() {
 
     // Hjelpeklasse
-    class ViewHolder(itemView: View):RecyclerView.ViewHolder(itemView){
+    class ViewHolder(itemView: View): RecyclerView.ViewHolder(itemView){
         private val itemViewTextView: TextView = itemView.findViewById(R.id.TodoItem_text)
+        private val itemViewCheckbox: CheckBox = itemView.findViewById(R.id.checkbox)
 
-        fun bind(word: String) { //fun bind(word: String, onItemClicked:(String) -> Unit) {
-            //binding.itemName.text = todoItem.ItemName // todoItem.itemName
+        fun bind(word: String, isChecked: Boolean) { //fun bind(word: String, onItemClicked:(String) -> Unit) {
             itemViewTextView.text = word
+            itemViewCheckbox.isChecked = isChecked
+            // ToDo: Legg til setOnClickListener
             /*itemViewTextView.setOnClickListener {
                 onItemClicked(String)
             }*/
@@ -39,9 +42,10 @@ class TodoItemRecyclerAdapter(
 
     // Display data at a certain position
     override fun onBindViewHolder(holder: ViewHolder, position: Int){
-        // ToDo: remove commment for bind, if this adapter is going to be used
-        //holder.bind(todoItems[position])
+        holder.bind(todoItems[position].itemName, todoItems[position].checked)
     }
 
     // ToDo: Legg til onClickListener() funksjonalitet for checkboxes
+
+    // ToDo: Funksjon for sletting av elementer
 }
